requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
dashscope==1.14.1
aiofiles==23.2.1
python-multipart==0.0.6
